rename: SignupMain as SignupMainStyle
