+++ src/components/pages/Signup/registration.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'next-i18next';
 import SignupLayout from '@/components/layout/SignupLayout';
 import StepHeader from './component/StepHeader';
 import { RegistrationLogoImageCss, StepLogoPositionCss } from './styles/RegistrationStyle';
-import { SignupMainContentCss, SignupMainNextButtonCss } from './styles/SignupMain';
+import { SignupMainContentCss, SignupMainNextButtonCss } from './styles/SignupMainStyle';
 
 const RegistrationPage: NextPageWithLayout = () => {
   const { t } = useTranslation(['page-signup'])
