+++ src/components/pages/Signup/styles/SignupMainStyle.ts
